refactor, cards responsabilities were separated from the game init logic, test for cards logic added
